Cyber/requirements_flask.txt
flask==2.3.2
flask-cors==4.0.0
plotly==5.17.0
pillow==10.1.0
cryptography==41.0.7
pycryptodome==3.19.0
joblib==1.3.2
psutil==5.9.6
werkzeug==2.3.6
jinja2==3.1.2
easyocr==1.7.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
